Refactor header code
